Control.Parallel.Eden.Edi

Description

The low-level parallel functional language: EDen Implementation language

This module defines a thin layer of type-checking wrappers over the parallel primitives implemented in ParPrim.hs.

Depends on the Eden Compiler.

Eden Group Marburg

Documentation

fork :: IO () -> IO ()

spawnProcessAt :: Int -> IO () -> IO ()

spawnArgsProcessAt :: NFData a => Int -> (a -> IO ()) -> a -> IO ()

data ChanName' a

createC :: IO (ChanName' a, a)

createCs :: Int -> IO ([ChanName' a], [a])

sendWith :: Strategy a -> ChanName' a -> a -> IO ()

sendNF :: NFData a => ChanName' a -> a -> IO ()

sendStreamWith :: (a -> ()) -> ChanName' [a] -> [a] -> IO ()

sendNFStream :: NFData a => ChanName' [a] -> [a] -> IO ()

noPe :: IO Int

selfPe :: IO Int

class NFData a where

A class of types that can be fully evaluated.

Methods

rnf :: a -> ()

rnf should reduce its argument to normal form (that is, fully evaluate all sub-components), and then return '()'.

The default implementation of rnf is

 rnf a = a `seq` ()

which may be convenient when defining instances for data types with no unevaluated fields (e.g. enumerations).

using :: a -> Strategy a -> a

Evaluate a value using the given strategy.

r0 :: Strategy a

r0 performs *no* evaluation.

rseq :: Strategy a

rseq evaluates its argument to weak head normal form.

rdeepseq :: NFData a => Strategy a

rdeepseq fully evaluates its argument. Relies on class NFData from module Control.DeepSeq.

seqList :: Strategy a -> Strategy [a]

Evaluate each element of a list according to the given strategy. This function is a specialisation of seqFoldable to lists.

seqFoldable :: Foldable t => Strategy a -> Strategy (t a)

Evaluate the elements of a foldable data structure according to the given strategy.